Moved out location service manager from nearby activity
