+++ app/src/main/java/fr/free/nrw/commons/location/LatLng.java
@@ -1,11 +1,11 @@
-package fr.free.nrw.commons.nearby;
+package fr.free.nrw.commons.location;
 
 public class LatLng {
 
     public final double latitude;
     public final double longitude;
 
-    LatLng(double latitude, double longitude) {
+    public LatLng(double latitude, double longitude) {
         if(-180.0D <= longitude && longitude < 180.0D) {
             this.longitude = longitude;
         } else {
