+++ app/src/main/java/fr/free/nrw/commons/nearby/Place.java
@@ -3,6 +3,8 @@ package fr.free.nrw.commons.nearby;
 import android.graphics.Bitmap;
 import android.net.Uri;
 
+import fr.free.nrw.commons.location.LatLng;
+
 public class Place {
 
     public String name;
